Prefer access return facts
Thread semantic-product receiver facts into call-return count and
access classification so stale local metadata cannot choose access
return kinds before graph-owned binding type IDs.

Add focused coverage for ID-first vector receiver inference and fix
compile-only blockers exposed by the backend IR target build.

+++ tests/unit/test_ir_pipeline_validation_ir_lowerer_statement_call_helper_validates_function_table_diagnostics.cpp
@@ -363,7 +363,8 @@ TEST_CASE("ir lowerer arithmetic helper uses semantic pointer operand facts befo
         if (it == localMap.end()) {
           return false;
         }
-        instructions.push_back({primec::IrOpcode::LoadLocal, it->second.index});
+        instructions.push_back(
+            {primec::IrOpcode::LoadLocal, static_cast<uint64_t>(it->second.index)});
         return true;
       },
       inferKind,
@@ -402,7 +403,8 @@ TEST_CASE("ir lowerer arithmetic helper uses semantic pointer operand facts befo
         if (it == localMap.end()) {
           return false;
         }
-        instructions.push_back({primec::IrOpcode::LoadLocal, it->second.index});
+        instructions.push_back(
+            {primec::IrOpcode::LoadLocal, static_cast<uint64_t>(it->second.index)});
         return true;
       },
       inferKind,
